Allow 32-bit registers in `classical.expr` system (#16449)
* Allow 32-bit registers in `classical.expr` system

The current limit in `Type` limits `Uint` sizes to 16-bit (65,535).  We
want to do scaling tests storing one million or more bits, so we need to
widen the type.  QPY already uses a 32-bit width in this location, and
the 16-bit internal limit was just a memory optimisation.

* Fix `printf` specifier


---------

+++ crates/cext/src/classical_expr.rs
@@ -101,7 +101,7 @@ pub struct CExprTypeInfo {
     /// The expression type
     ty: CExprType,
     /// Bit width for the Uint expression type
-    width: u16,
+    width: u32,
 }
 
 impl From<CExprTypeInfo> for Type {
